1. Separate tests for WorkerShim => created test-worker.html file 2. Use the approach suggested by @guybedford for detecting the `es-module-shims.js` path needed in WorkerShim constructor.
