+++ src/worker-shims.js
@@ -1,23 +1,15 @@
-import { baseUrl as pageBaseUrl, createBlob } from './common.js';
+import { baseUrl as pageBaseUrl, esModuleShimsSrc, createBlob } from './common.js';
 
 export class WorkerShim {
   constructor(aURL, options = {type: 'classic'}) {
-    if (options.type !== 'module') {
+    if (options.type !== 'module')
       return new Worker(aURL, options);
-    }
 
-    let es_module_shims_src = new URL('es-module-shims.js', pageBaseUrl).href;
-    const scripts = document.scripts;
-    for (let i = 0, len = scripts.length; i < len; i++) {
-      if (scripts[i].src.includes('es-module-shims.js')) {
-          es_module_shims_src = scripts[i].src;
+    if (!esModuleShimsSrc)
+      throw new Error('es-module-shims.js must be loaded with a script tag for WorkerShim support.');
 
-          break;
-      }
-    }
-
-    const workerScriptUrl = createBlob(`importScripts('${es_module_shims_src}'); 
-    self.importMap = ${JSON.stringify(options.importMap || {})}; importShim('${new URL(aURL, pageBaseUrl).href}')`);
+    const workerScriptUrl = createBlob(
+      `importScripts('${esModuleShimsSrc}'); self.importMap = ${JSON.stringify(options.importMap || {})}; importShim('${new URL(aURL, pageBaseUrl).href}')`);
 
     const workerOptions = {...options};
     workerOptions.type = 'classic';
