+++ test/test-worker.html
@@ -15,17 +15,13 @@
   }
 }
 </script>
-<script type="module-shim">
-  import test from "test";
-  console.log(test);
-</script>
-<script type="module" src="../src/es-module-shims.js"></script>
+<script type="text/javascript" src="../dist/es-module-shims.js"></script>
 <script type="module">
   mocha.setup('tdd');
   mocha.allowUncaught();
 
   self.baseURL = location.href.substr(0, location.href.lastIndexOf('/') + 1);
-  self.rootURL = location.href.substr(0, location.href.length - 'test/test.html'.length);
+  self.rootURL = location.href.substr(0, location.href.length - 'test/test-worker.html'.length);
   self.assert = function (val) {
     equal(!!val, true);
   };
@@ -39,7 +35,7 @@
     throw new Error(msg);
   };
 
-  const suites = ['browser-modules'];
+  const suites = ['worker'];
   function runNextSuite () {
     mocha.suite.suites.shift();
     const suite = suites.shift();
